requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.2
aiosignal==1.3.2
alpaca-py==0.40.1
alpaca-trade-api==3.2.0
annotated-types==0.7.0
async-timeout==5.0.1
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
contourpy==1.3.2
curl_cffi==0.11.1
cycler==0.12.1
deprecation==2.1.0
fonttools==4.58.1
frozendict==2.4.6
frozenlist==1.6.0
idna==3.10
kiwisolver==1.4.8
matplotlib==3.10.3
msgpack==1.0.3
multidict==6.4.4
multitasking==0.0.11
numpy==2.2.6
packaging==25.0
pandas==2.2.3
peewee==3.18.1
pillow==11.2.1
platformdirs==4.3.8
propcache==0.3.1
protobuf==6.31.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.1
requests==2.32.3
six==1.17.0
soupsieve==2.7
sseclient-py==1.8.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==1.26.20
websocket-client==1.8.0
websockets==10.4
yarl==1.20.0
